refactor(signals): promote core/signals/ → signals/ (top-level)
Signal is public API — its module path should reflect that.
Update 28 import references across widgets, composites, core capabilities,
cli, and bootstack/__init__.py. Remove signals from core/__init__.py imports.
