+++ src/bootstack/signals/integration.py
@@ -21,7 +21,7 @@ def _is_signal(obj: Any) -> bool:
         # Cheap duck-typing to avoid hard import dependency
         m = type(obj).__module__
         return (
-            (m.startswith("bootstack.signals") or m.startswith("bootstack.core.signals"))
+            (m.startswith("bootstack.signals") or m.startswith("bootstack.signals"))
             and hasattr(obj, "name")
             and hasattr(obj, "var")
         )
