+++ src/bootstack/signals/types.py
