+++ src/bootstack/widgets/composites/dropdownbutton.py
@@ -10,7 +10,7 @@ from bootstack.widgets.mixins import configure_delegate
 from bootstack.widgets.types import Master
 
 if TYPE_CHECKING:
-    from bootstack.core.signals import Signal
+    from bootstack.signals import Signal
 
 
 class DropdownButtonKwargs(TypedDict, total=False):
