+++ src/bootstack/widgets/composites/sidenav/item.py
@@ -14,7 +14,7 @@ from bootstack.widgets.mixins import configure_delegate
 from bootstack.widgets.types import Master
 
 if TYPE_CHECKING:
-    from bootstack.core.signals import Signal
+    from bootstack.signals import Signal
 
 
 class SideNavItemKwargs(TypedDict, total=False):
