+++ src/bootstack/widgets/mixins/signal_mixin.py
@@ -15,7 +15,7 @@ import tkinter as tk
 from typing import Any, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from bootstack.core.signals import Signal
+    from bootstack.signals import Signal
 
 from bootstack.core.capabilities.signals import (
     is_signal,
