+++ src/bootstack/widgets/primitives/button.py
@@ -11,7 +11,7 @@ from bootstack.widgets.mixins import IconMixin, TextSignalMixin, LocalizationMix
 from bootstack.widgets.types import Master
 
 if TYPE_CHECKING:
-    from bootstack.core.signals import Signal
+    from bootstack.signals import Signal
 
 
 class ButtonKwargs(TypedDict, total=False):
