+++ src/bootstack/widgets/primitives/combobox.py
@@ -12,7 +12,7 @@ from bootstack.widgets.types import Master
 from ..mixins import TextSignalMixin, configure_delegate
 
 if TYPE_CHECKING:
-    from bootstack.core.signals import Signal
+    from bootstack.signals import Signal
 
 
 class ComboboxKwargs(TypedDict, total=False):
