+++ src/bootstack/widgets/primitives/optionmenu.py
@@ -11,7 +11,7 @@ from bootstack.widgets.mixins import configure_delegate
 from bootstack.widgets.types import Master
 
 if TYPE_CHECKING:
-    from bootstack.core.signals import Signal
+    from bootstack.signals import Signal
 
 
 class OptionMenuChangeEventData(TypedDict):
